Remove filesystem dependency in config tests
Depending on the filesystem during unit tests is not too bad, but I'd
still consider it an integration test. It also slows down testing by a
random amount of time. I refactored config_handler to perform its logic
in a separately testable function and skipped the 'integration' test by
default.

I also rewrote the assertions in the original test to use jest's expect
instead of throwing errors. This is both more compact and provides
better output on failures.